requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudinary==1.25.0
Django==3.1.6
django-admin-rangefilter==0.7.0
django-cloudinary-storage==0.3.0
django-elasticsearch-dsl==7.1.4
elasticsearch==7.12.0
elasticsearch-dsl==7.3.0
idna==2.10
Pillow==8.1.2
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
